+++ homebridge-ui/public/js/main.js
@@ -22,6 +22,8 @@ function updateFormFromConfig() {
     document.getElementById('country').value = pluginConfig.country || "US";
     document.getElementById('countryInput1').value = pluginConfig.country || "US";
     document.getElementById('CameraMaxLivestreamDuration').value = pluginConfig.CameraMaxLivestreamDuration || 30;
+    document.getElementById('enableButton').checked = pluginConfig.enableButton || true;
+    document.getElementById('motionButton').checked = pluginConfig.motionButton || true;
     homebridge.fixScrollHeight();
 }
 
@@ -39,6 +41,8 @@ function updateConfigFromForm() {
     pluginConfig.ignoreDevices = document.getElementById('ignoreDevices').value.split(",");
     pluginConfig.country = document.getElementById('country').value;
     pluginConfig.CameraMaxLivestreamDuration = parseInt(document.getElementById('CameraMaxLivestreamDuration').value);
+    pluginConfig.enableButton = document.getElementById('enableButton').checked;
+    pluginConfig.motionButton = document.getElementById('motionButton').checked;
 }
 
 function adjustPollingValue() {